don't show report abuse for allowlisted games
